refactor: clean mod.rs files to canonical re-exports-only pattern
Split bots, favorites, search, preferences, and setup from single-file
modules with inline types/handlers/errors into proper multi-file layouts
(handlers.rs, types.rs, error.rs, validation.rs). Move admin Redis cache
helpers to shared.rs and auth hash_token() to helpers.rs. Update openapi.rs
path/schema references to the new submodule locations.
