+++ server/src/admin/mod.rs
@@ -19,65 +19,11 @@ use axum::middleware::from_fn_with_state;
 use axum::routing::{delete, get, post, put};
 use axum::Router;
 pub use error::AdminError;
-use fred::prelude::*;
 pub use middleware::{require_elevated, require_system_admin};
-use sqlx::PgPool;
+pub use shared::{cache_elevated_status, is_elevated_admin};
 pub use types::{ElevatedAdmin, SystemAdminUser};
-use uuid::Uuid;
 
 use crate::api::AppState;
-
-/// Check if a user is an elevated admin (for WebSocket subscription check).
-/// This checks both system admin status and elevated session validity.
-///
-/// Security: Always falls back to database on cache miss to ensure fail-secure behavior.
-pub async fn is_elevated_admin(redis: &Client, db: &PgPool, user_id: Uuid) -> bool {
-    // Check cache first (fast path)
-    let cache_key = format!("admin:elevated:{user_id}");
-    let cached: Option<String> = redis.get(&cache_key).await.ok().flatten();
-
-    if let Some(value) = cached {
-        return value == "1";
-    }
-
-    // Cache miss - fallback to database (fail-secure)
-    let is_elevated = check_elevated_in_db(db, user_id).await;
-
-    // Cache the result (60s TTL to balance freshness and load)
-    if is_elevated {
-        cache_elevated_status(redis, user_id, true, 60).await;
-    }
-
-    is_elevated
-}
-
-/// Check elevated session status directly in the database.
-async fn check_elevated_in_db(db: &PgPool, user_id: Uuid) -> bool {
-    queries::has_active_elevated_session(db, user_id)
-        .await
-        .unwrap_or(false)
-}
-
-/// Cache elevated admin status in Redis (called after elevation).
-pub async fn cache_elevated_status(
-    redis: &Client,
-    user_id: Uuid,
-    is_elevated: bool,
-    ttl_secs: i64,
-) {
-    let cache_key = format!("admin:elevated:{user_id}");
-    let value = if is_elevated { "1" } else { "0" };
-
-    let _: Result<(), _> = redis
-        .set(
-            &cache_key,
-            value,
-            Some(Expiration::EX(ttl_secs)),
-            None,
-            false,
-        )
-        .await;
-}
 
 /// Create the admin router.
 ///
